refactor(layering): extract the DataAccelerator contract into data-accelerator-api
Carve the accelerator **contract** out of `runtime`'s `dataaccelerator` module into
a new `data-accelerator-api` crate that sits below `runtime`. This is the seam that
lets the accelerator engines and the AcceleratedTable machinery depend on the
contract without depending on the orchestrator — a prerequisite for moving the
engines into `runtime-accelerator-*` crates (so `runtime` sheds the duckdb/rusqlite/
turso/tokio-postgres driver deps and starts compiling earlier) and for carving
`accelerated_table`.

The trait was already abstracted behind `&dyn AccelerationSource`, so no cycle-break
was needed. `create_accelerator_table` already takes the secrets handle as a
parameter and every dependency it names lives below runtime (`runtime-secrets`,
`runtime-parameters`, `runtime-table-partition`, datafusion + the dtp fork), so the
whole creation flow moves down as-is — callers (`registry.create_accelerator_table(...)`)
are unchanged.

Moved to `data-accelerator-api`:
- the `DataAccelerator` trait; `AcceleratorRegistration` + `DATA_ACCELERATOR_REGISTRATIONS`
  slice + the `register_data_accelerator!` macro; `AcceleratorEngineRegistry` and its
  `create_accelerator_table` flow; `AcceleratorExternalTableBuilder`;
  `acceleration_file_path`/`get_acceleration_layout`; the `Error`/`FilePathError`/
  `SnapshotReloadUnsupported`/`SchemaEvolutionUnsupported`/`ReloadProviderFactory` types;
  the `swappable` + `upsert_dedup` table-provider seams; and `spice_data_base_path`.

Stays in `runtime`: the engine implementations (arrow/duckdb/sqlite/turso/postgres/
cayenne) + `spice_sys`/`snapshots`/`storage`. `runtime` `pub use`-shims the moved items
so `crate::dataaccelerator::…` paths keep resolving; engines invoke the macro
path-qualified (`data_accelerator_api::register_data_accelerator!`).

- new crate: `[lints] workspace=true`, copyright header, layers.toml (shared-utility)
- layering guard green (132 crates, no upward edges)
