+++ crates/data-accelerator-api/src/swappable.rs
