+++ crates/data-accelerator-api/src/types.rs
@@ -19,7 +19,7 @@ limitations under the License.
 //! These types are re-exported from `dataaccelerator` so existing import paths
 //! remain unchanged.
 
-use crate::component::dataset::acceleration::Engine;
+use runtime_acceleration::Engine;
 use std::{collections::HashMap, sync::Arc};
 use tokio::sync::RwLock;
 
@@ -50,7 +50,7 @@ impl AcceleratorEngineRegistry {
         }
     }
 
-    pub(crate) async fn register_accelerator_engine(
+    pub async fn register_accelerator_engine(
         &self,
         engine: Engine,
         accelerator_engine: Arc<dyn DataAccelerator>,
@@ -67,7 +67,7 @@ impl AcceleratorEngineRegistry {
         }
     }
 
-    pub(crate) async fn register_all(&self) {
+    pub async fn register_all(&self) {
         for registration in DATA_ACCELERATOR_REGISTRATIONS {
             self.register_accelerator_engine(registration.engine, (registration.constructor)())
                 .await;
