+++ crates/runtime/Cargo.toml
@@ -43,6 +43,7 @@ chunking = { path = "../chunking" }
 clap.workspace = true
 csv = "1.3.0"
 dashmap = "6.1.0"
+data-accelerator-api = { path = "../data-accelerator-api" }
 data_components = { path = "../data_components", features = ["flightsql"] }
 dataformat-json = { path = "../dataformat-json" }
 datafusion.workspace = true
