+++ crates/runtime/src/dataaccelerator/arrow.rs
@@ -30,7 +30,6 @@ use std::{any::Any, sync::Arc};
 
 use crate::component::dataset::acceleration::{Engine, RefreshMode};
 use crate::parameters::ParameterSpec;
-use crate::register_data_accelerator;
 
 use super::{AccelerationSource, DataAccelerator};
 
@@ -143,4 +142,4 @@ impl DataAccelerator for ArrowAccelerator {
     }
 }
 
-register_data_accelerator!(Engine::Arrow, ArrowAccelerator);
+data_accelerator_api::register_data_accelerator!(Engine::Arrow, ArrowAccelerator);
