+++ layers.toml
@@ -200,7 +200,7 @@ yaml = "foundation"
 # --- shared-utility: composable libraries the runtime is built from (default
 # tier, but listed here for clarity / auditability). Everything else defaults to
 # `shared-utility`:
-#   app, cache, cayenne, data_components, datafusion-ddl,
+#   app, cache, cayenne, data-accelerator-api, data_components, datafusion-ddl,
 #   datafusion-optimizer-rules, llms, package, runtime-acceleration,
 #   runtime-auth, runtime-cluster, runtime-datafusion, runtime-datafusion-udfs,
 #   runtime-parameters, runtime-request-context, runtime-search, runtime-secrets,
